wpt_babel/content-security-policy/inheritance/window.html

<!DOCTYPE html>
<script src="/resources/testharness.js">
</script>
<script src="/resources/testharnessreport.js">
</script>
<meta content="img-src 'none'" http-equiv="Content-Security-Policy"/>
<body>
 <script>
  "use strict";

function wait_for_error_from_window(w, test) {
  window.addEventListener('message', test.step_func(function (e) {
    if (e.source != w) return;
    assert_equals(e.data, "error");
    w.close();
    test.done();
  }));
}
async_test(function (t) {
  var w = window.open();
  var img = document.createElement('img');
  img.onerror = t.step_func_done(function (_) {
    return w.close();
  });
  img.onload = t.unreached_func();
  img.src = "/images/red-16x16.png";
  w.document.body.appendChild(img);
}, "window.open() inherits policy.");
async_test(function (t) {
  var w = window.open();
  wait_for_error_from_window(w, t);
  w.document.write("\n      <img src='/images/red-16x16.png'\n        onload='window.opener.postMessage(\"load\", \"*\");'\n        onerror='window.opener.postMessage(\"error\", \"*\");'\n      >\n    ");
}, "`document.write` into `window.open()` inherits policy.");
async_test(function (t) {
  var b = new Blob(["\n        <img src='".concat(window.origin, "/images/red-16x16.png'\n          onload='window.opener.postMessage(\"load\", \"*\");'\n          onerror='window.opener.postMessage(\"error\", \"*\");'\n        >\n      ")], {
    type: "text/html"
  });
  wait_for_error_from_window(window.open(URL.createObjectURL(b)), t);
}, "window.open('blob:...') inherits policy.");

// Navigation to top-level `data:` is blocked.

async_test(function (t) {
  var url = "javascript:\"<img src='".concat(window.origin, "/images/red-16x16.png'\n          onload='window.opener.postMessage(\\\"load\\\", \\\"*\\\");'\n          onerror='window.opener.postMessage(\\\"error\\\", \\\"*\\\");'\n        >\"");
  wait_for_error_from_window(window.open(url), t);
}, "window.open('javascript:...') inherits policy.");
 </script>
</body>
